add GrpcXdsBootstrapBuilder, move HTTP filter registry creation there
